Recursive makefile structure and refactoring.
