+++ freeze_thaw.cpp
@@ -3,7 +3,7 @@
 //*******************************Inclusions*******************************//
 //************************************************************************//
 #include "disk.h"
-#include "Vector.h"
+#include "Vector.cpp"
 #include<iostream>
 #include<fstream>
 #include<cstdlib>
@@ -251,5 +251,5 @@ int main(int argc, char *argv[])//Main program
 //************************************************************************//
 //********************************Functions*******************************//
 //************************************************************************//
-#include "helper.hpp"
+#include "helper.h"
