+++ makefile.inc
@@ -5,4 +5,10 @@ RM = rm
 SHELL = /bin/sh
 ECHO = echo
 
+CFLAGS = -c -Wall
+LFLAGS = -Wall
+INCS = -I./
+OBJS = disk.o Vector.o
+EXEC = freezethaw
+
 .SILENT :
