+++ source/disk.cpp
@@ -1,4 +1,4 @@
-#include "disk.h"
+#include "disk.hpp"
 #include <stdlib.h>
 
 //Gears Integration Scheme - Predict
